1.3log_mqtt/requirement.txt
Django==4.2.7
paho-mqtt==1.6.1